Use analyze to find plaintext secret in audit mode
In audit mode, we were using secret_generator to find the
plaintext secret. This was problematic because it led to
inconsistences between what the scan functionality would
find and what the audit functionality would find, leading
to user errors.

Using the same plugin function for both scan and audit
will lead to fewer user errors.