Handle binary values specially in yaml file parser
This is step (1) in supporting binary in both YAML and non-
YAML files.

This makes it so that instead of immediately converting the
base64-encoded binary into a binary value in python, we just
interpret the binary as a normal string, but annotate it as
such with the `is_binary` flag.

This is needed so that plugins can scan a different value
from the value hashed into baselines.