Detect YAML binary secrets
This implements support for high-entropy secrets in binary
values in yaml files.

We encode the binary value into a hex- or base64-encoded
string (based on the plugin), and run the normal entropy
check. If the string is deemed to be high-entropy, we re
encode the string into a yaml binary (using `yaml.dump`)
and strip the `!!binary`. This yaml binary is considered
the secret, and is put into the baseline as normal.

I had to update a test function so that it uses a custom
hex high-entropy detector, since `HighEntropyStringsPlugin`
is now an abstract class.