better info logs
With this change, debug logs are going to be a *lot* more verbose, by
logging every time a line is scanned. In doing so, users are able to see
the line and filename that causes their scan to fail.

As a consequence, previous debug logs will be upgraded to "info" logs.