Fix 'no plugins found' bug on python > 3.7
Fixes the 'no plugins found' bug on python > 3.7 by explicitly passing
the current settings object to an initialization function that runs
before the multiprocessing job is started. This initialization creates a
new settings object for that process and sets its properties equal to
that of the settings object passed from the parent process.